use the correct format specifier for NSInteger. Might have resolved chrome cast button disappearing problem.

+++ Externals/TouchUI/Geometry.m
@@ -129,7 +129,7 @@ CGRect ScaleAndAlignRectToRect(CGRect inImageRect, CGRect inDestinationRect, EIm
 
 NSString *NSStringFromCIntegerPoint(CIntegerPoint inPoint)
 {
-    return([NSString stringWithFormat:@"%d,%d", inPoint.x, inPoint.y]);
+    return([NSString stringWithFormat:@"%ld,%ld", (long)inPoint.x, (long)inPoint.y]);
 }
 
 extern CIntegerPoint CIntegerPointFromString(NSString *inString)
